Refactor UpdateProfileFrom and integrate with EditProfilePage
